+++ src/__fixtures__/destructuring-properties/invalid-comments.ts
