+++ src/__fixtures__/destructuring-properties/valid-comments.ts
