+++ src/__tests__/imports.spec.ts
@@ -1,7 +1,7 @@
 jest.mock("../resolver")
 
 import { RuleTester } from "eslint"
-import rule, { SortGroup } from "../rules/sort-imports"
+import rule, { SortGroup } from "../rules/imports"
 import { invalidFixture, validFixture } from "./utils"
 
 const messages = rule.meta!.messages! as Record<
