+++ src/rules/destructuring-properties.ts
