+++ src/rules/object-properties.ts
